requirements.txt

Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
Pillow>=9.0
django-environ>=0.9.0
djangorestframework-gis>=0.17
django-filter>=23.1
gunicorn>=20.1.0
